requirements.txt
fastapi
uvicorn
pydantic
motor
pyjwt
passlib[bcrypt]
python-jose
